+++ study-app/src/app/login/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import AuthLayout from '../../components/common/AuthLayout';
-import LoginForm from '../../components/loginPage/LoginForm';
+import LoginForm from '../../components/login/LoginForm';
 
 const LoginPage = (): React.ReactNode => {
   return (
