+++ study-app/src/app/signup/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import AuthLayout from '../../components/common/AuthLayout';
-import SignUpForm from '../../components/signUpPage/SignUpForm';
+import SignUpForm from '../../components/signUp/SignUpForm';
 
 const SignUpPage = (): React.ReactNode => {
   return (
