+++ study-app/src/components/landing/ImageSlider.tsx
